Add SchemaInferenceStack
Migrating away from use of SchemaPath requires us to introduce a
concept which can be used to transfer inference state across functional
blocks.

SchemaInferenceStack is precisely such a concept. It is inherently
attacked to an EffectiveModelContext and can be manipulated by users
to reflect traversal through EffectiveStatements. The statements can
be pushed and popped based on application decisions, but the stack
always retains information as to how we got to a particular statement.

This provides superior accuracy when compared to SchemaPath-driven
operation, which is ambiguous around groupings. It also allows
efficient transfer of this state for the purposes of further inference
-- for example to perform static analysis of a leaf-ref type.

JIRA: YANGTOOLS-1086
Change-Id: Id50a2daac0af6c5944b4bb2ef29734f7efb5315b
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>