Add DataTreeCandidates.aggregate() method
MD-SAL datastore access patterns allow for state compression to
occur, as changes are reported in terms of an ordered collection
of DataTreeCandidate objects.

These typically carry all state transitions, which an application
may (or may not) be interested in. In case an application does not
care about seeing all transitions, it may be beneficial to compress
these state transitions, so that they appear to have been caused
by a single transaction.

In order to achieve that, this patch introduces a new utility
method, DataTreeCandidates.aggregate(), which aggregates such a
collecting into a single DataTreeCandidate -- hence allowing
infrastructure components to perform state compression behind
application's back.

JIRA: YANGTOOLS-1014
Change-Id: I31094e303dbbb8430cb427ae6c59ee7956e0117c
Signed-off-by: tadei.bilan <tadei.bilan@pantheon.tech>