Switch SchemaAwareApplyOperation error reporting
Switch to reporting IllegalStateException when the target schema
node is not as opposed to IllegalArgumentException. This makes it
clear that we have entered internal inconsistency rather than
the user doing something wrong.

Furthermore eliminate the possibility of throwing an unchecked
exception when the user has made a wrong reference -- opting for
a well-controlled checked exception instead.

JIRA: YANGTOOLS-1105
Change-Id: Icd3ddb03a721358d07427e1fddeec8a71e73b933
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>