Add EffectiveStatementEquivalent
Our implementation classes provide both EffectiveStatement world
as well as SchemaNode world view in a single object. As we are
making the transition we need to formalize this contract, so
downstreams can operate on both old and new views.

To make this easier, we mark vast majority of constructs with a
new interface, EffectiveStatementEquivalent, which allows users
to get EffectiveStatement view of a particular SchemaNode. With
this interface users can provide a single EffectiveStatement-based
implementation and provide a convenience method, or directly ask
their users to perform the conversion.

A few statements are not transitioned yet, as their mapping is
ambiguous.

JIRA: YANGTOOLS-1119
Change-Id: I8aa6627912498faf8d88873f7db08fe0ab524adf
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>