Add support for regex-related OpenConfig extensions
There are two OpenConfig extensions, which augment how regular
expressions are handled.

The "regexp-posix" extension instructs the parser to interpret any
pattern arguments as not XSD regular expressions, but rather as POSIX
extended expressions.

The "posix-pattern" extension acts as a proper carrier of the POSIX
expression, so that "pattern" statements have correct XSD patterns.

// FIXME: PatternStatementSupport support needs to understand that
          "regexp-posix" is in effect
// FIXME: add a test suite

JIRA: YANGTOOLS-1132
Change-Id: I80a713c1b5dca051adbbb55720ae22535aeac9d6
Signed-off-by: Martin Bobak <martin.bobak@pantheon.tech>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>