Add OpenConfig pattern-related statement definitions
OpenConfig defines two extensions to deal with pattern statements.

The first is a semantic marker to tell the parser that OpenConfig's
interpretation of what is 'pattern' is used within that source file,
i.e. patterns are some variant of POSIX regular expression.

The second one is a properly-defined YANG extension which carries
the POSIX RE. This extension is a sibling to pattern statement,
whose argument is a YANG-compliant XSD regular expression.

JIRA: YANGTOOLS-1132
Change-Id: I6bd1c074eda2a02eea59913c939b9c236fdf8e8c
Signed-off-by: Martin Bobak <martin.bobak@pantheon.tech>