Convert yang-xpath-impl into a JPMS module
This is a bit more involved, as we are using a different
javax.inject and do not use osgi.cmpn but
org.osgi.service.component.annotations directly.

The latter results in a dependency on an automatic module, but
that is not really a problem as we never touch them at runtime.

JIRA: YANGTOOLS-1145
Change-Id: Ie80b02eba551fa89c334f20484e386d808316ca3
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>