WIP: Add Global{Identifier,Statement}Namespace
Globally-unique instances are special enough to warrant some attention
and special treatment. The reason for that is we do a lot of global
lookups and each of them requires a walk towards the global storage
node, which is always BuildGlobalContext.

This should help speeding these up, the NamespaceStorageNode interface
hierarchy now knows when it is looking for BuildGlobalContext. While
it is falling back to parent lookups, this strengthens the case for
having root statement available quickly, as the two lookups are
very similar.

We might take advantage of that once we reign in the number of storage
nodes we allocate, which is subject of YANGTOOLS-874.

// FIXME: document all conversions (most are)
// FIXME: finish audit of callers
// FIXME: rename globalOf() back to global()

JIRA: YANGTOOLS-1159
Change-Id: I3cac8b3769e2432bc4708273ffa989959adc7c6c
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>