Perform partial substatement initialization
When we have a request for a schema tree participant, we should not
force materialization of

We should not need to initialize all substatements when we are being
asked for a single one.

Side-step this need by keeping a separate map of statements we have
initialized and meshing it together if need be.

JIRA: YANGTOOLS-1160
Change-Id: I8e014650272409ba30e528341dcba9c05311971b
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>