Reformulate StatementContextFactory.createEffective()
We need a better handle on what state is being accessed during
createEffective() invocation.

This is achieved by isolating this state into EffectiveStmtCtx,
which gets passed to StatementFactory.createEffective() along
with single-use streams of declared and effective statements.

This interface is then implemented by a proxy in front of a
StatementContextBase, providing effective isolation.

There are a few places which still require direct access to
the underlying StmtContext, but those are limited and subject
to elimination. All of them access backing context through
Current.cearnbannog(), making it easy to spot.

Fix false checks of augmentation substatements. AugmentTest
checks if augment with no when statement contains case
which is not true but it gave us such result before.

JIRA: YANGTOOLS-1161
Change-Id: I950c466d02c75be1e21a9d0606b990d35e4ca5c2
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>