Reformulate StatementContextFactory.createEffective()
We need a better handle on what state is being accessed during
createEffective() invocation.

This is achieved by isolating this state into EffectiveStmtCtx,
which gets passed to StatementFactory.createEffective() along
with single-use streams of declared and effective statements.

This interface is then implemented by a proxy in front of a
StatementContextBase, providing effective isolation.

There are a few places which still require direct access to
the underlying StmtContext, but those are limited and subject
to elimination. All of them access backing context through
Current.cearnbannog(), making it easy to spot.

// FIXME: revert augment tests and removal

JIRA: YANGTOOLS-1161
Change-Id: I950c466d02c75be1e21a9d0606b990d35e4ca5c2
Signed-off-by: miroslav.kovac <miroslav.kovac@pantheon.tech>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>