Add CommonStmtCtx
The idea here is that EffectiveStmtCtx represents an effective view,
whereas StmtContext is an inference view. They share some bits at
the very least.

This patch splits out the common stateless bit, which is not affected
by any sort of interpretation on our part into CommonStmtCtx. This
information is always available and valid.

JIRA: YANGTOOLS-1185
Change-Id: Ia852195811bbd3ec0091fa54488ada230e7f6245
Signed-off-by: miroslav.kovac <miroslav.kovac@pantheon.tech>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>