WIP: Do not retain substatements in InferredStatementContext
We really do not want to retain substatements once a statement
has been built.

This is a trivial attempt at poisoning effectiveStatements() once
we have built the target.

// FIXME: this does not really work, as it breaks when we have
//        Inferred(A) with delegate being Inferred(B). If B ends
//        up being built first, we disallow access and end up
//        exploding when A attempts to initialize its substatements.

The point here is that InferredStatementContext.createSubstatements()
must not touch upstream for prototype.mutableEffectiveSubstatements()
if it is an InferredStatementContext:

We should have recorded the proper state by forcing prototype to
give us something between Materialized and ForwardingSubstatements,
something which eventually blows up to be a full Materialized.

JIRA: YANGTOOLS-1161
Change-Id: Id5935ac43977c83dafa97154456474f445e1321c
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>