WIP: Do not retain substatements in InferredStatementContext
We really do not want to retain substatements once a statement
has been built.

The problem we are solving is that the copy's build operation
may occur after the source's build operation, hence we still need
access to substatements.

We take the approach of having three basic states and checking
their transitions. When a statement is copied, we acquire source's
future materialization in the copy and perform access through there.

//
// FIXME: and this does not really work
//

There is caerbannog().allSubstatements() access, which is just not
valid. We'll deal with that separately.

The trouble is that the copy operation itself may be occuring after
the effective statement has been built, hence we never had a chance
to hook onto source before the substatements were thrown away.

This yet again feels like we need a new ModelProcessingPhase, and
this time it is real, as we need a recursive inference pass over
all statements to clean up these substatements.

So whenever inference decides to create a copy (augment, uses, etc.),
it also needs to create an inference action which prevents
ModelProcessingPhase.CLEANUP from completing.

When a statement copy has acquired source's effective substatements,
it will clear the action, eventually allowing it to be cleaned up.

JIRA: YANGTOOLS-1184
Change-Id: Id5935ac43977c83dafa97154456474f445e1321c
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>