Add a refcount mechanism for substatements
We need to defer cleaning substatements until there is no way
they could be referenced again. This can happen, for example,
when we make a copy through uses: source grouping can be built
independently, but needs to retain its effective statements
until all its copies are built as well.

We fuse the two remaining booleans in StatementContextBase to
allow up to 16 bits for refcount, which should be enough for
most scenarios.

// FIXME: RootStatementContext's namespaces

JIRA: YANGTOOLS-1184
Change-Id: I5d26846ee1a5ce705671d9e66d45137a3d1084cf
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>