Isolate statement reuse sites
Augment wants to do something atypical: it assumes its state cannot
be further inferred into at both source and target context place and
it wants the same EffectiveStatement to emerge at both places.

This violates SchemaPath rules on one side of the equation or the
other, but seems to work for the intended case.

As it turns out, we are already doing something similar in the reactor,
but here we are taking the guidance of StatementSupport in the form
of CopyPolicy.

Make sure all users of this mechanics are calling into
Mutable.replicaAsChildOf(), where the reactor can properly capture
the lifecycle transition. This includes 'deviate' statement as well.

JIRA: YANGTOOLS-1190
Change-Id: Ic0032c50a1abbc65265dcbc038f72598761e9c36
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>