Make sweep-on-decrement more aggressive
We are holding off on cleaning if we have a parent, which effectively
means we not start sweeping children until that process is initiated
from parent.

This is needlessly conservative: we should actually be checking parent
refcount and trigger cleanup as soon as we observe no explicit
references.

JIRA: YANGTOOLS-1184
Change-Id: Ief30aa03b760116540d77038fb8c78fd5b0daf5a
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>