WIP: Use correct key-arg splitting
The splitter we have operates on spaces and trims results to get
rid of whitespace -- probably as a consequence of us historically
having bugs there.

We currently break if someone uses double-quoted whitespace trimming,
as we do not treat '\n' as a separator and then attempt to interpret
it as a node-identifier.

Add an explicit SEP equivalent CharMatcher and use a Splitter on that,
resulting in '\t' and '\n' being correctly trimmed.

Change-Id: Ifa3085fffcbbe24204e9d6c0d86ed8c41bd61065
JIRA: YANGTOOLS-1200
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>