Reimplement SchemaContextUtil.getBaseTypeForLeafRef()
This is the primary entry point to resolution of TypeDefinitions
and generally inferring along the data tree. We already have
required utilities in SchemaInferenceStack.

Introduce SchemaInferenceStack.resolvePathExpression(), which
serves the serves the same purpose, but results in a stack
manipulation.

This completely removes the need to perform any string-based
operations, or dedicated SchemaNode lookups or anything of that
kind -- removing a lot of legacy crud.

// FIXME: mold I76425a6ed4c6d7648124086a490082e9189a8d1c into this patch

JIRA: YANGTOOLS-1127
JIRA: YANGTOOLS-1229
Change-Id: I64e4f4ef2062ab2a0d143a520a53e8972cf54b26
Signed-off-by: miroslav.kovac <miroslav.kovac@pantheon.tech>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>